% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{binomial choose}
\usage{
bin_choose(k, n)
}
\arguments{
\item{k}{is the number of successes; is an integer}

\item{n}{is the number of trials; is an integer}

\item{prob}{a probability value}
}
\value{
the number of calculations in which k  successes can occur in n trials
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
}
