% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials; is an integer}

\item{prob}{probability of success; is a probability}
}
\value{
the kurtosis of the binomial
}
\description{
Calculates the kurtosis of the binomal distribution
}
\examples{
# 10 trials with probability of success 0.5
bin_kurtosis(10, 0.5)

}
