% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials; is an integer}

\item{prob}{probability of success; is a probability}
}
\value{
the variance of the binomial
}
\description{
Calculates the variance of the binomal distribution
}
\examples{
# 10 trials with probability of success 0.5
bin_variance(10, 0.5)

}
